requirements.txt
numpy
geopandas
contextily
matplotlib
PyQt6
geopy
